requirements.txt
#run pip install -r requirements.txt

gunicorn
selenium~=4.25.0
requests~=2.32.3
docker~=7.1.0
flask~=3.0.3
Werkzeug~=3.0.4
python-dotenv~=1.0.1
pydantic~=2.10.4
pyodbc
beautifulsoup4